backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-socketio==5.10.0
authlib==1.2.1
python-dotenv==1.0.0
loguru==0.7.2
slowapi==0.1.9
pydantic-settings==2.10.1
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
email-validator==2.1.0
httpx==0.27.0
authlib==1.3.0